Fix docker availability handling accross platforms (#139199)
* Fix docker availability handling accross platforms
    - Invert logic in isExcludedOs to correctly apply exclusions only in CI.
    - Fail tests in CI if Docker is missing on supported OS instead of skipping.
    - Update assumption messages for clarity.
